[cli] extract substrate rpc module
